@@ -1,15 +1,20 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 	<PropertyGroup>
 		<OutputType>Exe</OutputType>
-		<TargetFramework>net8.0</TargetFramework>
+		<TargetFramework>net9.0</TargetFramework>
 		<ImplicitUsings>enable</ImplicitUsings>
 		<Nullable>enable</Nullable>
 		<ValidlyConfiguration>
 			<ExitEarly>disabled</ExitEarly>
 		</ValidlyConfiguration>
+		<OptimizationPreference>speed</OptimizationPreference>
+		<IsAotCompatible Condition="$([MSBuild]::IsTargetFrameworkCompatible('$(TargetFramework)', 'net7.0'))"
+			>true</IsAotCompatible
+		>
+		<!-- <RuntimeIdentifier>win-x64</RuntimeIdentifier> -->
 	</PropertyGroup>
 	<ItemGroup>
-		<PackageReference Include="BenchmarkDotNet" Version="0.14.0" />
+		<PackageReference Include="BenchmarkDotNet" Version="0.15.4" />
 	</ItemGroup>
 	<ItemGroup>
 		<ProjectReference Include="..\..\Validly\Validly.csproj" />
