@@ -1,3 +1,3 @@
 namespace Validly.SourceGenerator.Utils.Mapping;
 
-public sealed record DependencyInjectionInfo(string Name, bool IsKeyedService, object? Key);
+public record DependencyInjectionInfo(string Name);
