@@ -116,7 +116,9 @@ internal static class ValidatableObjectIncrementalValueProvider
 	{
 		if (
 			validatableAttribute
-				.NamedArguments.FirstOrDefault(static x => x.Key == nameof(ValidatableAttribute.UseAutoValidators))
+				.NamedArguments.FirstOrDefault(static x =>
+					x.Key == Consts.ValidatableAttributeUseAutoValidatorsPropertyName
+				)
 				.Value.Value
 			is true
 		)
@@ -126,7 +128,9 @@ internal static class ValidatableObjectIncrementalValueProvider
 
 		if (
 			validatableAttribute
-				.NamedArguments.FirstOrDefault(static x => x.Key == nameof(ValidatableAttribute.NoAutoValidators))
+				.NamedArguments.FirstOrDefault(static x =>
+					x.Key == Consts.ValidatableAttributeNoAutoValidatorsPropertyName
+				)
 				.Value.Value
 			is true
 		)
@@ -141,7 +145,7 @@ internal static class ValidatableObjectIncrementalValueProvider
 	{
 		if (
 			validatableAttribute
-				.NamedArguments.FirstOrDefault(static x => x.Key == nameof(ValidatableAttribute.UseExitEarly))
+				.NamedArguments.FirstOrDefault(static x => x.Key == Consts.ValidatableAttributeUseExitEarlyPropertyName)
 				.Value.Value
 			is true
 		)
@@ -151,7 +155,7 @@ internal static class ValidatableObjectIncrementalValueProvider
 
 		if (
 			validatableAttribute
-				.NamedArguments.FirstOrDefault(static x => x.Key == nameof(ValidatableAttribute.NoExitEarly))
+				.NamedArguments.FirstOrDefault(static x => x.Key == Consts.ValidatableAttributeNoAExitEarlyPropertyName)
 				.Value.Value
 			is true
 		)
