@@ -8,13 +8,15 @@ internal static class StaticKeys
 	public enum KeysEnum
 	{
 		Key1,
-		Key2
+		Key2,
 	}
 
 	public const KeysEnum EnumKey = KeysEnum.Key1;
 	public const string StringKey = "StringKey";
+	public const char CharKey = 'A';
 	public const string UnknownStringKey = "UnknownStringKey";
 	public const int IntKey = 123;
+	public const bool BoolKey = true;
 
 	public static IValidatable? GetValidatable(object key)
 	{
@@ -24,7 +26,9 @@ internal static class StaticKeys
 			string => new ValidatableObjectStringKeyed(stringKeyValue),
 			int => new ValidatableObjectIntKeyed(stringKeyValue),
 			KeysEnum => new ValidatableObjectEnumKeyed(stringKeyValue),
-			_ => null
+			bool => new ValidatableObjectBoolKeyed(stringKeyValue),
+			char => new ValidatableObjectCharKeyed(stringKeyValue),
+			_ => null,
 		};
 	}
 }
@@ -33,13 +37,15 @@ internal static class StaticKeys
 internal partial record ValidatableObjectEnumKeyed([property: CustomValidation] string Property)
 {
 	public IEnumerable<ValidationMessage> ValidateProperty(
-		[FromKeyedServices(StaticKeys.EnumKey)]
-		IDependency? dependency)
+		[FromKeyedServices(StaticKeys.EnumKey)] IDependency? dependency
+	)
 	{
 		if (dependency is null)
 		{
-			yield return new ValidationMessage($"Dependency cannot be null for property {Property}",
-				"Dependency.NotNull");
+			yield return new ValidationMessage(
+				$"Dependency cannot be null for property {Property}",
+				"Dependency.NotNull"
+			);
 		}
 	}
 }
@@ -48,12 +54,15 @@ internal partial record ValidatableObjectEnumKeyed([property: CustomValidation] 
 internal partial record ValidatableObjectIntKeyed([property: CustomValidation] string Property)
 {
 	public IEnumerable<ValidationMessage> ValidateProperty(
-		[FromKeyedServices(StaticKeys.IntKey)] IDependency? dependency)
+		[FromKeyedServices(StaticKeys.IntKey)] IDependency? dependency
+	)
 	{
 		if (dependency is null)
 		{
-			yield return new ValidationMessage($"Dependency cannot be null for property {Property}",
-				"Dependency.NotNull");
+			yield return new ValidationMessage(
+				$"Dependency cannot be null for property {Property}",
+				"Dependency.NotNull"
+			);
 		}
 	}
 }
@@ -62,13 +71,49 @@ internal partial record ValidatableObjectIntKeyed([property: CustomValidation] s
 internal partial record ValidatableObjectStringKeyed([property: CustomValidation] string Property)
 {
 	public IEnumerable<ValidationMessage> ValidateProperty(
-		[FromKeyedServices(StaticKeys.StringKey)]
-		IDependency? dependency)
+		[FromKeyedServices(StaticKeys.StringKey)] IDependency? dependency
+	)
 	{
 		if (dependency is null)
 		{
-			yield return new ValidationMessage($"Dependency cannot be null for property {Property}",
-				"Dependency.NotNull");
+			yield return new ValidationMessage(
+				$"Dependency cannot be null for property {Property}",
+				"Dependency.NotNull"
+			);
+		}
+	}
+}
+
+[Validatable]
+internal partial record ValidatableObjectCharKeyed([property: CustomValidation] string Property)
+{
+	public IEnumerable<ValidationMessage> ValidateProperty(
+		[FromKeyedServices(StaticKeys.CharKey)] IDependency? dependency
+	)
+	{
+		if (dependency is null)
+		{
+			yield return new ValidationMessage(
+				$"Dependency cannot be null for property {Property}",
+				"Dependency.NotNull"
+			);
+		}
+	}
+}
+
+[Validatable]
+internal partial record ValidatableObjectBoolKeyed([property: CustomValidation] string Property)
+{
+	public IEnumerable<ValidationMessage> ValidateProperty(
+		[FromKeyedServices(StaticKeys.BoolKey)] IDependency? dependency
+	)
+	{
+		if (dependency is null)
+		{
+			yield return new ValidationMessage(
+				$"Dependency cannot be null for property {Property}",
+				"Dependency.NotNull"
+			);
 		}
 	}
 }
