@@ -7,44 +7,45 @@ namespace Validly.Tests.ServiceProvider;
 public class ServiceProviderHelperTests
 {
 	[Theory]
-	[InlineData("Property")]
-	[InlineData("Not Property")]
+	[InlineData("Odd")]
+	[InlineData("Even")]
 	public void Validate_DependencyReturnedValue_ResultChanged(string property)
 	{
 		var dependency = Substitute.For<IDependency>();
-		dependency.GetNumber(Arg.Any<string>()).Returns(property is "Property" ? 1 : 2);
+		dependency.GetNumber(Arg.Any<string>()).Returns(property is "Odd" ? 1 : 2);
+		var serviceProvider = new ServiceCollection().AddSingleton(dependency).BuildServiceProvider();
 
-		var validatable = new ValidatableObject("Property");
-		var result = validatable.Validate(new ServiceCollection().AddSingleton(dependency).BuildServiceProvider());
+		var validatable = new ValidatableObject("Odd");
+		var result = validatable.Validate(serviceProvider);
 
-		Assert.Equal(result.IsSuccess, property is "Property");
+		Assert.Equal(result.IsSuccess, property is "Odd");
 	}
 
 	[Theory]
 	[InlineData(StaticKeys.StringKey)]
 	[InlineData(StaticKeys.IntKey)]
 	[InlineData(StaticKeys.EnumKey)]
+	[InlineData(StaticKeys.BoolKey)]
+	[InlineData(StaticKeys.CharKey)]
 	[InlineData(StaticKeys.UnknownStringKey, false)]
 	public async Task Validate_DependencyReturnedValue_KeyReturnedValue(object key, bool shouldSucceed = true)
 	{
 		var dependency = Substitute.For<IDependency>();
 		dependency.GetNumber(Arg.Any<string>()).Returns(1);
-		var serviceCollection = new ServiceCollection()
-			.AddKeyedSingleton(key, dependency)
-			.BuildServiceProvider();
+		var serviceProvider = new ServiceCollection().AddKeyedSingleton(key, dependency).BuildServiceProvider();
 
 		var validatable = StaticKeys.GetValidatable(key);
 
 		if (shouldSucceed)
 		{
-			var result =
-				await validatable!.ValidateAsync(serviceCollection);
+			var result = await validatable!.ValidateAsync(serviceProvider);
 			Assert.Equal(shouldSucceed, result.IsSuccess);
 		}
 		else
 		{
 			await Assert.ThrowsAsync<InvalidOperationException>(async () =>
-				await validatable!.ValidateAsync(serviceCollection));
+				await validatable!.ValidateAsync(serviceProvider)
+			);
 		}
 	}
 }
