@@ -7,7 +7,7 @@ namespace Validly;
 /// It serves as a marker for a source generator to know that the class is validatable.
 /// </remarks>
 [AttributeUsage(AttributeTargets.Class)]
-public class ValidatableAttribute : Attribute
+public sealed class ValidatableAttribute : Attribute
 {
 	/// <summary>
 	/// Enables auto validators for the class
