@@ -4,4 +4,4 @@ namespace Validly.Validators;
 /// Marks a class as a validator
 /// </summary>
 [AttributeUsage(AttributeTargets.Class)]
-public class ValidatorAttribute : Attribute;
+public sealed class ValidatorAttribute : Attribute;
